requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.30.6
streamlit==1.38.0
requests==2.32.3
openai==1.51.0
pinecone==5.3.0
langchain==0.3.1
langchain_openai==0.2.1
langchain_pinecone==0.2.0
python-dotenv==1.0.1